backend/requirements.txt
# Environment Variables
python-dotenv==1.0.0

# HTTP Requests (for routing service and AI API)
requests==2.31.0

# AI Integration (AIPIPE/OpenAI compatible)
openai>=1.0.0

# HTTP Server
flask==3.0.0
flask-cors
flask-login
flask-bcrypt

# MongoDB Database
pymongo==4.6.1
dnspython==2.4.2

# SMS Service (Twilio)
twilio>=8.0.0
